AuthServer/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
djangogrpcframework==0.2.1
djangorestframework==3.14.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-api-keys==0.5.5
googleapis-common-protos==1.62.0
grpc-interceptor==0.15.4
grpcio==1.60.0
grpcio-status==1.60.0
grpcio-tools==1.60.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
